+++ tests/time_integration.rs
@@ -38,3 +38,21 @@ fn supports_ago_since_for_offset_datetimes() {
         "1h 1m ago"
     );
 }
+
+#[cfg(feature = "polish")]
+#[test]
+fn supports_locale_aware_ago_since_for_offset_datetimes() {
+    let then = ::time::OffsetDateTime::from_unix_timestamp(0).unwrap();
+    let now = ::time::OffsetDateTime::from_unix_timestamp(3665).unwrap();
+    let out = humtime::ago_since_with(
+        then,
+        now,
+        DurationOptions::new()
+            .locale(humfmt::locale::Polish)
+            .long_units()
+            .max_units(3),
+    )
+    .unwrap();
+
+    assert_eq!(out.to_string(), "1 godzina 1 minuta 5 sekund temu");
+}